python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
uagents==0.22.10
httpx==0.25.2
pydantic==2.5.0
anthropic==0.71.0

